# backend/docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/artists/{artistId}/pfp" : {
      "put" : {
        "tags" : [ "artist-controller" ],
        "summary" : "Replace an artist's profile picture",
        "description" : "Only the artist's creator. Multipart field `file`: same rules as the user profile picture. Returns the new public URL.",
        "operationId" : "updateArtistPfp",
        "parameters" : [ {
          "name" : "artistId",
          "in" : "path",
          "description" : "Public id of the artist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "a_8f3k2Q"
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary",
                    "description" : "Image file, JPEG or PNG"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfilePictureUpdate"
                }
              }
            }
          },
          "400" : {
            "description" : "Missing `file` part (code REQUEST_FAILED), or empty, unreadable or too small image (codes INVALID_IMAGE, IMAGE_TOO_SMALL)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller didn't create the artist (code NOT_ARTIST_CREATOR); or Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "404" : {
            "description" : "Artist not found (code ARTIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "413" : {
            "description" : "Over 5MB or 25 megapixels (code IMAGE_TOO_LARGE), or file over 25MB (code REQUEST_FAILED)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "415" : {
            "description" : "Not a JPEG or PNG (code UNSUPPORTED_IMAGE_TYPE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 20 requests per minute per IP, shared by all upload endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/artists/{artistId}/favorite" : {
      "put" : {
        "tags" : [ "artist-controller" ],
        "summary" : "Favorite an artist",
        "description" : "Does nothing if the caller already favorited the artist.",
        "operationId" : "favoriteArtist",
        "parameters" : [ {
          "name" : "artistId",
          "in" : "path",
          "description" : "Public id of the artist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "a_8f3k2Q"
        } ],
        "responses" : {
          "204" : {
            "description" : "Artist is now a favorite"
          },
          "404" : {
            "description" : "Artist not found (code ARTIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "403" : {
            "description" : "Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "artist-controller" ],
        "summary" : "Unfavorite an artist",
        "description" : "Does nothing if the artist wasn't a favorite or doesn't exist.",
        "operationId" : "unfavoriteArtist",
        "parameters" : [ {
          "name" : "artistId",
          "in" : "path",
          "description" : "Public id of the artist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "a_8f3k2Q"
        } ],
        "responses" : {
          "204" : {
            "description" : "Artist is no longer a favorite"
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "403" : {
            "description" : "Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/artists/{artistId}/banner" : {
      "put" : {
        "tags" : [ "artist-controller" ],
        "summary" : "Replace an artist's banner image",
        "description" : "Only the artist's creator. Multipart fields `file`, `cropX`, `cropY`, `cropWidth`, `cropHeight`:\nsame rules as the user banner. Returns the new public URL.\n",
        "operationId" : "updateArtistBanner",
        "parameters" : [ {
          "name" : "artistId",
          "in" : "path",
          "description" : "Public id of the artist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "a_8f3k2Q"
        }, {
          "name" : "cropX",
          "in" : "query",
          "description" : "Left edge of the crop area, in pixels of the original image",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "cropY",
          "in" : "query",
          "description" : "Top edge of the crop area, in pixels of the original image",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "cropWidth",
          "in" : "query",
          "description" : "Width of the crop area, in pixels",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "cropHeight",
          "in" : "query",
          "description" : "Height of the crop area, in pixels",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary",
                    "description" : "Image file, JPEG or PNG"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BannerUpdate"
                }
              }
            }
          },
          "400" : {
            "description" : "Missing, unreadable or too small image, or missing or bad crop (codes INVALID_IMAGE, IMAGE_TOO_SMALL, INVALID_CROP, REQUEST_FAILED)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller didn't create the artist (code NOT_ARTIST_CREATOR); or Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "404" : {
            "description" : "Artist not found (code ARTIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "413" : {
            "description" : "Over 5MB or 25 megapixels (code IMAGE_TOO_LARGE), or file over 25MB (code REQUEST_FAILED)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "415" : {
            "description" : "Not a JPEG or PNG (code UNSUPPORTED_IMAGE_TYPE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 20 requests per minute per IP, shared by all upload endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/artists/{artistId}" : {
      "get" : {
        "tags" : [ "artist-controller" ],
        "summary" : "Get an artist",
        "description" : "Public. `isFavorite` is true when the caller has favorited the artist. `canEdit` is true when the\ncaller created the artist and may change its pfp and banner. Both are false with no token.\n`artistPfpUrl` and `artistBannerUrl` are null until set.\n",
        "operationId" : "getArtist",
        "parameters" : [ {
          "name" : "artistId",
          "in" : "path",
          "description" : "Public id of the artist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "a_8f3k2Q"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArtistDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Artist not found (code ARTIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        }, { } ]
      }
    },
    "/api/artists/favorites/{username}" : {
      "get" : {
        "tags" : [ "artist-controller" ],
        "summary" : "List a user's favorite artists",
        "description" : "Public. Newest favorite first.",
        "operationId" : "getFavoriteArtists",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "Username of the user whose favorites to list",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "justin"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ArtistSummary"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "User not found (code USER_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ProfilePictureUpdate" : {
        "type" : "object",
        "description" : "The new profile picture",
        "properties" : {
          "profilePictureUrl" : {
            "type" : "string",
            "description" : "Public URL of the new WebP picture",
            "example" : "https://aux.justinabruinsma.com/pfp/3f2b8c1e-8d4a-4c6e-9a51-2f0d7b9e6c11.webp"
          }
        }
      },
...
      "BannerUpdate" : {
        "type" : "object",
        "description" : "The new banner",
        "properties" : {
          "bannerUrl" : {
            "type" : "string",
            "description" : "Public URL of the new WebP banner",
            "example" : "https://aux.justinabruinsma.com/banner/3f2b8c1e-8d4a-4c6e-9a51-2f0d7b9e6c11.webp"
          }
        }
      },
...
      "ArtistSummary" : {
        "type" : "object",
        "description" : "The short form of an artist, shown on cards and next to music pieces",
        "properties" : {
          "artistId" : {
            "type" : "string",
            "description" : "Public id of the artist",
            "example" : "a_8f3k2Q"
          },
          "artistName" : {
            "type" : "string",
            "description" : "Artist name",
            "example" : "Tame Impala"
          },
          "artistPfpUrl" : {
            "type" : "string",
            "description" : "Artist profile picture URL; null until set",
            "example" : "https://aux.justinabruinsma.com/artist-pfp/3f2b8c1e-8d4a-4c6e-9a51-2f0d7b9e6c11.webp"
          }
        }
      },
...
      "AuxServerError" : {
        "description" : "Body of every 4xx and 5xx response",
        "properties" : {
          "errorDetails" : {
            "$ref" : "#/components/schemas/ErrorDetails"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Token from /api/auth/login or /api/auth/register",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}